harmonize auth barrel + rename Admin to SuperAdmin
- rename Roles.Admin to Roles.SuperAdmin ('super-admin')
- add modules/auth/index.{ts,tsx} barrel for react/vue/nuxt re-exporting Auth + entities + Roles_Auth + Roles
- generated app file uses 'import { Auth } from ./modules/auth' across all frameworks
- drop authEntities object from auth barrel exports (export entities individually)
- simplify sveltekit override: keep only the post-copy Roles.ts removal
