+++ projects/create-remult/templates/auth/better-auth/react/src/modules/auth/index.tsx
@@ -1,0 +1,9 @@
+export { default as Auth } from "./Auth";
+export {
+  User,
+  Session,
+  Account,
+  Verification,
+  Roles_Auth,
+} from "./authEntities";
+export { Roles } from "./Roles";
