+++ projects/create-remult/templates/auth/better-auth/shared/src/modules/auth/Roles.ts
@@ -2,6 +2,6 @@ import { Roles_Auth } from "./authEntities.js";
 
 /** ALL ROLES of your application. [Learn more](https://remult.dev/docs/modules#roles) */
 export const Roles = {
-  Admin: "admin",
+  SuperAdmin: "super-admin",
   ...Roles_Auth,
 } as const;
