+++ projects/create-remult/templates/auth/better-auth/sveltekit/src/lib/index.ts
@@ -2,6 +2,6 @@ import { Roles_Auth } from "$modules/auth";
 
 /** ALL ROLES of your application. [Learn more](https://remult.dev/docs/modules#roles) */
 export const Roles = {
-  Admin: "admin",
+  SuperAdmin: "super-admin",
   ...Roles_Auth,
 } as const;
